services/orchestrator/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
jinja2==3.1.4
pydantic==2.9.2
requests==2.32.3
python-multipart==0.0.9
playwright==1.46.0
pdfplumber==0.11.4
python-docx==1.1.2
python-pptx==0.6.23
numpy==2.1.1
PyYAML==6.0.2
